requirements.txt
pandas
numpy
streamlit
fast-autocomplete[levenshtein]
qrcode
fuzzywuzzy
sentence_transformers
nltk
scikit_learn
spacy
python-Levenshtein
python-docx
matplotlib
wordcloud